refactor(datagrid): collapse cell hierarchy to single DataGridCellView
