--- TablePro/Views/Results/Cells/DataGridBlobCellView.swift
@@ -1,12 +1,0 @@
-//
-//  DataGridBlobCellView.swift
-//  TablePro
-//
-
-import AppKit
-
-final class DataGridBlobCellView: DataGridChevronCellView {
-    override class var reuseIdentifier: NSUserInterfaceItemIdentifier {
-        NSUserInterfaceItemIdentifier("dataCell.blob")
-    }
-}
