--- TablePro/Views/Results/Cells/DataGridBooleanCellView.swift
@@ -1,12 +1,0 @@
-//
-//  DataGridBooleanCellView.swift
-//  TablePro
-//
-
-import AppKit
-
-final class DataGridBooleanCellView: DataGridChevronCellView {
-    override class var reuseIdentifier: NSUserInterfaceItemIdentifier {
-        NSUserInterfaceItemIdentifier("dataCell.boolean")
-    }
-}
