--- TablePro/Views/Results/Cells/DataGridDateCellView.swift
@@ -1,12 +1,0 @@
-//
-//  DataGridDateCellView.swift
-//  TablePro
-//
-
-import AppKit
-
-final class DataGridDateCellView: DataGridChevronCellView {
-    override class var reuseIdentifier: NSUserInterfaceItemIdentifier {
-        NSUserInterfaceItemIdentifier("dataCell.date")
-    }
-}
