--- TablePro/Views/Results/Cells/DataGridDropdownCellView.swift
@@ -1,12 +1,0 @@
-//
-//  DataGridDropdownCellView.swift
-//  TablePro
-//
-
-import AppKit
-
-final class DataGridDropdownCellView: DataGridChevronCellView {
-    override class var reuseIdentifier: NSUserInterfaceItemIdentifier {
-        NSUserInterfaceItemIdentifier("dataCell.dropdown")
-    }
-}
