--- TablePro/Views/Results/Cells/DataGridJsonCellView.swift
@@ -1,12 +1,0 @@
-//
-//  DataGridJsonCellView.swift
-//  TablePro
-//
-
-import AppKit
-
-final class DataGridJsonCellView: DataGridChevronCellView {
-    override class var reuseIdentifier: NSUserInterfaceItemIdentifier {
-        NSUserInterfaceItemIdentifier("dataCell.json")
-    }
-}
