--- TablePro/Views/Results/Cells/DataGridTextCellView.swift
@@ -1,12 +1,0 @@
-//
-//  DataGridTextCellView.swift
-//  TablePro
-//
-
-import AppKit
-
-final class DataGridTextCellView: DataGridBaseCellView {
-    override class var reuseIdentifier: NSUserInterfaceItemIdentifier {
-        NSUserInterfaceItemIdentifier("dataCell.text")
-    }
-}
